add disponible & progressbar column

+++ Models/DTOs/RubroResponseDTO.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public int RubroTypeId { get; set; }
         public string tipoRubroNombre { get; set; }
-        public int valorInicial { get; set; }
-        public int valorGastado { get; set; }
+        public decimal valorInicial { get; set; }
+        public decimal valorGastado { get; set; }
         public int Mes { get; set; }
         public int Anio { get; set; }
     }
